Add files via upload
In the C programming language, we were asked to build a solution to the constrained buffer
the producer-consumer dilemma utilizing shared memory and fork/exec routines. So we created The
producer file, which forks a consumer process and replaces the child process with the appropriate
consumer code using exec (), Following that, we write test cases to run the code in various
scenarios to check if there is an issue.